Add files via upload
This study aims to provide an in-depth understanding of the topic of central government debt in the European Union (EU). The first section, the Literature Review, is dedicated to an analysis of current knowledge. Here, we check previous studies, theoretical stances and empirical results to establish a basis for investigation.

The second part reveals the Theoretical Framework, providing the basis for our analysis. This part investigates the main theories and economic models relevant to public debt, prepares the analysisfor empirical exploration.

In the third part, the Methodology, we elucidate the research pattern, data sources and tools used to analyze the relationships and dynamics of central government debt in the EU. This section serves as a methodological basis that guides the study.

In the fourth part we analyze the Results obtained from the Fractional Component Models - Logit. Here, the knowledge obtained from the models offers the relationships between independent variables and central government debt over the period from 1997 to 2020.

The fifth part of the study presents the findings of the study. Here, we reflect on the results, discuss their significance for policymakers, and propose potential ways for future research.